backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
python-multipart==0.0.22
Pillow==12.1.1
pillow-heif==1.3.0
aiofiles==25.1.0
openai>=1.30.0
python-jose[cryptography]==3.3.0
bcrypt==4.3.0
